requirements.txt
pandas
numpy
torch
antspyx
torchvision
torchaudio
monai
tensorboard
tqdm
matplotlib
wandb